Controller based process flow

+++ dicom2cloud/clientgui.py
@@ -5,7 +5,7 @@ import time
 from glob import iglob
 from hashlib import sha256
 from os import R_OK, W_OK, mkdir, access, walk
-from os.path import join, isdir, split, exists
+from os.path import join, isdir, split, exists, basename
 
 import pydicom as dicom
 from pydicom.errors import InvalidDicomError
@@ -77,19 +77,17 @@ class HomePanel(WelcomePanel):
 ########################################################################
 class ProcessRunPanel(ProcessPanel):
     def __init__(self, parent):
-
         super(ProcessRunPanel, self).__init__(parent)
-        self.db = DBI()
-        choices = self.db.getCaptions()
-        self.m_checkListProcess.AppendItems(choices)
         # Set up event handler for any worker thread results
         EVT_RESULT(self, self.progressfunc)
         # EVT_CANCEL(self, self.stopfunc)
         # Set timer handler
         self.start = {}
         self.toggleval = 0
-        self.server = ''
+        #self.server = ''
         self.controller = Controller()
+        choices = self.controller.db.getCaptions()
+        self.m_checkListProcess.AppendItems(choices)
 
     def progressfunc(self, msg):
         """
@@ -157,7 +155,7 @@ class ProcessRunPanel(ProcessPanel):
         event.Skip()
 
     def OnShowDescription(self, event):
-        desc = self.db.getDescription(event.String)
+        desc = self.controller.db.getDescription(event.String)
         self.m_stTitle.SetLabelText(event.String)
         self.m_stDescription.SetLabelText(desc)
         self.Layout()
@@ -184,10 +182,10 @@ class ProcessRunPanel(ProcessPanel):
         btn = event.GetEventObject()
         btn.Disable()
         # Get selected processes
-        selection = self.m_checkListProcess.GetCheckedItems()
-        print("Processes selected: ", selection)
+        processes = self.m_checkListProcess.GetCheckedStrings()
+        print("Processes selected: ", processes)
         # Get Cloud Server
-        self.server = self.m_server.GetStringSelection().lower()
+        server = self.m_server.GetStringSelection().lower()
         # Get Output directory and File list
         filepanel = self.getFilePanel()
         if filepanel.outputdir is not None:
@@ -205,10 +203,10 @@ class ProcessRunPanel(ProcessPanel):
             raise IOError(msg)
 
         num_files = filepanel.m_dataViewListCtrl1.GetItemCount()
-        print('Selected Files:', num_files)
+        print('Selected Series:', num_files)
 
         try:
-            if len(selection) > 0 and num_files > 0:
+            if len(processes) > 0 and num_files > 0:
                 # Add files to db
                 #
                 # csvheader = ['Filename', 'Series', 'Process', 'Server']
@@ -216,12 +214,26 @@ class ProcessRunPanel(ProcessPanel):
                 #     writer = csv.DictWriter(csvfile, fieldnames=csvheader)
                 #     writer.writeheader()
 
-                for i in range(0, num_files):
-                    if filepanel.m_dataViewListCtrl1.GetToggleValue(i, 0):
-                        seriesid = filepanel.m_dataViewListCtrl1.GetValue(i, 6)
-                        # for each series, create temp dir and copy files
-                        self.m_stOutputlog.SetLabelText("Copying DICOM data %s ... please wait" % seriesid)
-                        dest = self.copyseries(seriesid, targetdir)
+                # for i in range(0, num_files):
+                #     if filepanel.m_dataViewListCtrl1.GetToggleValue(i, 0):
+                #         seriesid = filepanel.m_dataViewListCtrl1.GetValue(i, 6)
+                #         # for each series, create temp dir and copy files
+                #         self.m_stOutputlog.SetLabelText("Copying DICOM data %s ... please wait" % seriesid)
+                #         (dest,uuid) = self.copyseries(seriesid, targetdir)
+
+                # For each process, own thread per fileset
+                row = 0
+                for p in processes:
+                    for i in range(0, num_files):
+                        if filepanel.m_dataViewListCtrl1.GetToggleValue(i, 0):
+                            seriesid = filepanel.m_dataViewListCtrl1.GetValue(i, 6)
+                            # for each series, create temp dir and copy files
+                            self.m_stOutputlog.SetLabelText("Copying DICOM data %s ... please wait" % seriesid)
+                            uuid = self.generateuid(seriesid)
+                            if self.copyseries(uuid, targetdir):
+                                self.m_stOutputlog.SetLabelText("Process thread started %s ... please wait" % seriesid)
+                                self.controller.RunProcess(self, targetdir, seriesid, p, server, row)
+                                row = row + 1
                             #
                             # # TODO: Call Docker with series (dest) - then poll
                             # t = DockerThread(self, dest, seriesid, selection, self.server)
@@ -231,7 +243,7 @@ class ProcessRunPanel(ProcessPanel):
                             #     {'Filename': dest, 'Series': seriesid, 'Process': selection, 'Server': self.server})
 
             else:
-                if len(selection) == 0:
+                if len(processes) == 0:
                     msg = "No processes selected"
                 elif filepanel.outputdir is None:  # or len(filepanel.outputdir)<=0:
                     msg = "No outputdir provided"
@@ -244,24 +256,24 @@ class ProcessRunPanel(ProcessPanel):
             # Enable Run button
             self.m_btnRunProcess.Enable()
 
-    def copyseries(self, seriesnum, targetdir):
+    def copyseries(self, suuid, targetdir):
         """
-        Copy series to temp dir and save location
+        Copy series to temp dir if doesn't exist and save location
         :param self:
         :param seriesnum:
         :param targetdir:
-        :return:
+        :return: False if no files to copy or destination folder with copied files
         """
-        suuid = self.generateuid(seriesnum)
-        seriesfiles = self.db.getFiles(suuid)
-        dest = None
+        seriesfiles = self.controller.db.getFiles(suuid)
+        dest = False
         if seriesfiles is not None and len(seriesfiles) > 0:
             #subdir = self.generateuid(seriesnum)
             dest = join(targetdir, suuid)
             if not exists(dest):
                 mkdir(dest)
             for f in seriesfiles:
-                shutil.copy(f, dest)
+                if not exists(join(dest,basename(f))):
+                    shutil.copy(f, dest)
         return dest
 
     def generateuid(self, seriesnum):
@@ -347,20 +359,22 @@ class FileSelectPanel(FilesPanel):
                 continue
 
             # Check DICOM header info
-
             series_num = str(dcm.SeriesInstanceUID)
             uuid = self.generateuid(series_num)
             imagetype = str(dcm.ImageType[2])
             dicomdata = {'uuid': uuid,
                          'patientid': str(dcm.PatientID),
                          'patientname': str(dcm.PatientName),
-                         'series_num': series_num,
+                         'seriesnum': series_num,
                          'sequence': str(dcm.SequenceName),
                          'protocol': str(dcm.ProtocolName),
                          'imagetype': imagetype
                          }
-            self.db.addDicomdata(dicomdata)
-            self.db.addDicomfile(uuid, filename)
+
+            if not self.db.hasUuid(uuid):
+                self.db.addDicomdata(dicomdata)
+            if not self.db.hasFile(uuid,filename):
+                self.db.addDicomfile(uuid, filename)
 
             # if series_num not in global_series:
             #     global_series[series_num] = {'dicomdata': dicomdata, 'files': []}
@@ -381,7 +395,7 @@ class FileSelectPanel(FilesPanel):
                  self.db.getDicomdata(uuid,'sequence'),
                  self.db.getDicomdata(uuid,'protocol'),
                  self.db.getDicomdata(uuid,'imagetype'), str(numfiles),
-                 self.db.getDicomdata(uuid,'series_num')])
+                 self.db.getDicomdata(uuid,'seriesnum')])
             # #for s0 in global_series.items():
             # s = s0[1]['dicomdata']
             # numfiles = len(s0[1]['files'])
@@ -403,6 +417,64 @@ class FileSelectPanel(FilesPanel):
     def OnClearlist(self, event):
         print("Clear items in list")
         self.m_dataViewListCtrl1.DeleteAllItems()
+
+########################################################################
+class CloudRunPanel(CloudPanel):
+    def __init__(self, parent):
+        super(CloudRunPanel, self).__init__(parent)
+
+    def OnUpdate(self, event):
+        """
+        Load dummydatabase and for each seriesID - poll class
+        :param event:
+        :return:
+        """
+        filepanel = self.getFilePanel()
+        self.outputdir = filepanel.outputdir
+        dbfile = join(self.outputdir, 'dummydatabase.txt')
+        csvheader = ['Filename', 'Series', 'Process', 'Server']
+
+        self.m_tcResults.AppendText("\n***********\nCloud processing results\n***********\n")
+        with open(dbfile) as csvfile:
+            reader = csv.DictReader(csvfile)
+            for row in reader:
+                print(row['Filename'], row['Server'])
+                seriesid = split(row['Filename'])[1]
+                server = row['Server'].lower()
+                # Get uploader class and query
+                uploaderClass = get_class(server)
+                uploader = uploaderClass(seriesid)
+                done = uploader.isDone()
+                if done:
+                    uploader.download(join(self.outputdir, seriesid, 'download.tar'))
+                    msg = 'Series: %s \n\tSTATUS: Complete (%s)\n' % (
+                        seriesid, join(self.outputdir, seriesid, 'download.tar'))
+
+                else:
+                    msg = 'Series: %s \n\tSTATUS: Still processing\n' % seriesid
+                self.m_tcResults.AppendText(msg)
+        print('Finished cloud panel update')
+
+    def getFilePanel(self):
+        """
+        Get access to filepanel
+        :return:
+        """
+        filepanel = None
+
+        for fp in self.Parent.Children:
+            if isinstance(fp, FileSelectPanel):
+                filepanel = fp
+                break
+        return filepanel
+
+    def OnClearOutput(self, event):
+        """
+        Clear output panel
+        :param event:
+        :return:
+        """
+        self.m_tcResults.Clear()
 
 
 ########################################################################
@@ -483,65 +555,6 @@ class AppMain(wx.Listbook):
             self.Destroy()
         else:
             e.Veto()
-
-
-########################################################################
-class CloudRunPanel(CloudPanel):
-    def __init__(self, parent):
-        super(CloudRunPanel, self).__init__(parent)
-
-    def OnUpdate(self, event):
-        """
-        Load dummydatabase and for each seriesID - poll class
-        :param event:
-        :return:
-        """
-        filepanel = self.getFilePanel()
-        self.outputdir = filepanel.outputdir
-        dbfile = join(self.outputdir, 'dummydatabase.txt')
-        csvheader = ['Filename', 'Series', 'Process', 'Server']
-
-        self.m_tcResults.AppendText("\n***********\nCloud processing results\n***********\n")
-        with open(dbfile) as csvfile:
-            reader = csv.DictReader(csvfile)
-            for row in reader:
-                print(row['Filename'], row['Server'])
-                seriesid = split(row['Filename'])[1]
-                server = row['Server'].lower()
-                # Get uploader class and query
-                uploaderClass = get_class(server)
-                uploader = uploaderClass(seriesid)
-                done = uploader.isDone()
-                if done:
-                    uploader.download(join(self.outputdir, seriesid, 'download.tar'))
-                    msg = 'Series: %s \n\tSTATUS: Complete (%s)\n' % (
-                    seriesid, join(self.outputdir, seriesid, 'download.tar'))
-
-                else:
-                    msg = 'Series: %s \n\tSTATUS: Still processing\n' % seriesid
-                self.m_tcResults.AppendText(msg)
-        print('Finished cloud panel update')
-
-    def getFilePanel(self):
-        """
-        Get access to filepanel
-        :return:
-        """
-        filepanel = None
-
-        for fp in self.Parent.Children:
-            if isinstance(fp, FileSelectPanel):
-                filepanel = fp
-                break
-        return filepanel
-
-    def OnClearOutput(self, event):
-        """
-        Clear output panel
-        :param event:
-        :return:
-        """
-        self.m_tcResults.Clear()
 
 
 ########################################################################
